Extract `useColumnSizing` and `useRowSelection`
