+++ packages/ui/react-ui-table/src/hooks/index.ts
@@ -2,5 +2,7 @@
 // Copyright 2024 DXOS.org
 //
 
+export * from './useColumnSizing';
 export * from './useColumnSorting';
 export * from './usePinLastRow';
+export * from './useRowSelection';
